Fix strage animal sniffer issue related to nestmates
