+++ apm-agent-plugins/apm-jdk-httpclient-plugin/src/main/java/co/elastic/apm/agent/httpclient/HttpRequestHeadersInstrumentation.java
@@ -18,19 +18,10 @@
  */
 package co.elastic.apm.agent.httpclient;
 
-import co.elastic.apm.agent.impl.transaction.Span;
-import co.elastic.apm.agent.sdk.advice.AssignTo;
-import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.NamedElement;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
-
-import javax.annotation.Nullable;
-import java.net.http.HttpHeaders;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
 
 import static net.bytebuddy.matcher.ElementMatchers.hasSuperType;
 import static net.bytebuddy.matcher.ElementMatchers.nameContains;
@@ -54,18 +45,8 @@ public class HttpRequestHeadersInstrumentation extends AbstractHttpClientInstrum
         return named("headers").and(returns(named("java.net.http.HttpHeaders")));
     }
 
-    public static class AdviceClass {
-        @Nullable
-        @AssignTo.Return
-        @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class, inline = false)
-        public static HttpHeaders onAfterExecute(@Advice.Return @Nullable final HttpHeaders httpHeaders) {
-            Span span = tracer.getActiveSpan();
-            if (span == null || httpHeaders == null) { // in case of thrown exception return value might be null
-                return httpHeaders;
-            }
-            Map<String, List<String>> headersMap = new LinkedHashMap<>(httpHeaders.map());
-            span.propagateTraceContext(headersMap, HttpClientRequestPropertyAccessor.instance());
-            return HttpHeaders.of(headersMap, (x, y) -> true);
-        }
+    @Override
+    public String getAdviceClassName() {
+        return "co.elastic.apm.agent.httpclient.HttpRequestHeadersAdvice";
     }
 }
